requirements.txt
requests
python-jenkins>=1.4.0